clean up compile warnings across libgag, src, libusl
Mixed bag of -Wunused, -Woverloaded-virtual, -Wreorder /
use-before-init, unused-private-field, and unused-include diagnostics
across 46 files. Behavior-preserving cleanup pass — no logic changes.

Notable items:
- FertilityCalculatorDialog: rename no-arg execute() to runModal() to
  drop the name collision with OverlayScreen::execute(gfx, stepLength)
  and the using-declaration that masked it.
- libusl/native.h: NativeValuePrototype passes nullptr to Prototype()
  instead of the not-yet-constructed inherited 'heap' member; heap is
  fixed up by NativeValue's ctor.
- libgag/GUITabScreen: drop dead 'first'/'last' locals in onTimer,
  fall back to windows.begin()->first directly after removeGroup.

+++ src/MultiplayerGameScreen.cpp
@@ -153,11 +153,11 @@ void MultiplayerGameScreen::onAction(Widget *source, Action action, int par1, in
 			game->leaveGame();
 			endExecute(Cancelled);
 		}
-		else if ((par1 >= ADD_AI) && (par1 < ADD_AI + AI::SIZE))
+		else if ((par1 >= ADD_AI) && (par1 < ADD_AI + static_cast<int>(AI::SIZE)))
 		{
 			game->addAIPlayer((AI::ImplementitionID)(par1-ADD_AI));
 		}
-		else if ((par1>=CLOSE_BUTTONS)&&(par1<CLOSE_BUTTONS+MAX_NUMBER_OF_PLAYERS))
+		else if ((par1>=CLOSE_BUTTONS)&&(par1<static_cast<int>(CLOSE_BUTTONS)+MAX_NUMBER_OF_PLAYERS))
 		{
 			game->kickPlayer(par1 - CLOSE_BUTTONS);
 		}
